local/vision/requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
boto3>=1.26.0
python-dotenv>=0.19.0
pyzmq>=24.0.0
